Prodigy/task5/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
pillow>=9.5.0
opencv-python>=4.8.0.76
matplotlib>=3.7.0
tqdm>=4.66.0



